+++ resources/views/auditoria/ver_auditoria.blade.php
@@ -13,8 +13,9 @@
 			{{Breadcrumbs::render('auditoria')}}
 
 		<!-- The Modal -->
+		@if ($permiso->crear == 1)
 		<button type="button" onclick="window.location='{{ route("auditoria.create") }}'" class="btn btn-info ink-reaction btn-primary addbutton" id="myBtn"><span class="fa fa-plus"></span></button>
-
+		@endif
 
 		@endsection()
 
@@ -34,6 +35,7 @@
 					</thead>
 					<tbody>
 						@foreach ($audiorias as $audioria)
+						@if ($permiso->consultar == 1)
 						<tr>
 							<td>{{$audioria->Codigo}}</td>
 						@if($audioria->EmpresaAudita != '' || $audioria->EmpresaAudita != null)
@@ -45,6 +47,7 @@
 							<td>{{$audioria->NombreAuditoria}}</td>
 
 							<td>
+							@if ($permiso->eliminar == 1)
 								<div class="col-sm-6">
 
 									{!! Form::open(['route' => ['auditoria.destroy', $audioria->IdAuditoria], 'method' => 'DELETE']) !!}
@@ -53,16 +56,18 @@
 
 									{!! Form::close() !!}
 								</div>
-
+							@endif
+							@if ($permiso->actualizar == 1)
 								<div class="col-sm-6">
 
 									<a href="{{route('auditoria.edit', $audioria->IdAuditoria) }}" class="btn btn-primary btn-block editbutton" ><div class="gui-icon"><i class="fa fa-pencil"></i></div></a>
 
 								</div>
-
+							@endif
 							</td>
 							{{-- <td>{{$ata->Activo}}</td> --}}
 						</tr>
+						@endif
 						@endforeach
 					</tbody>
 				</table>
