+++ resources/views/auditoria/ver_informe_inspeccion.blade.php
@@ -12,7 +12,9 @@
 		@section('card-title')
 		{{ Breadcrumbs::render('informeinspeccion') }}
 		<!-- Begin Modal -->
+		@if ($permiso->crear == 1)
 		<button type="button" onclick="window.location='{{ route("informeInspeccion.create") }}'" class="btn btn-info ink-reaction btn-primary addbutton" id="myBtn"><span class="fa fa-plus"></span></button>
+		@endif
 		{{-- End modal --}}
 
 
@@ -34,11 +36,13 @@
 					</thead>
 					<tbody>
 						@foreach ($informesInspeccion as $informeInspeccion)
+						@if ($permiso->consultar == 1)
 						<tr>
 							<td>{{$informeInspeccion->Codigo}}</td>
 							<td>{{$informeInspeccion->FechaInicio}}</td>
 							<td>{{$informeInspeccion->NombreTipo}}</td>
 							<td>
+							@if ($permiso->eliminar == 1)
 								<div class="col-sm-6">
 
 									{!! Form::open(['route' => ['informeInspeccion.destroy', $informeInspeccion->IdCrearInforme], 'method' => 'DELETE']) !!}
@@ -47,16 +51,18 @@
 
 									{!! Form::close() !!}
 								</div>
-
-
+							@endif
+							@if ($permiso->actualizar == 1)
 								<div class="col-sm-6">
 
 									<a href="{{route('informeInspeccion.edit', $informeInspeccion->IdCrearInforme) }}" class="btn btn-primary btn-block editbutton" ><div class="gui-icon"><i class="fa fa-pencil"></i></div></a>
 
 								</div>
+							@endif
 							</td>
 							{{-- <td>{{$ata->Activo}}</td> --}}
 						</tr>
+						@endif
 						@endforeach
 					</tbody>
 				</table>
