+++ resources/views/fomento/empresas/ver_tablas_empresas.blade.php
@@ -12,12 +12,11 @@ Tablas Crear Empresa
 @section('card-title')
 {{ Breadcrumbs::render('empresa') }}
 
-@foreach ($perfil as $itemPerfil)
-	@if ($itemPerfil->IdRol == 12)
+
 	<!-- The Modal -->
+	@if ($permiso->crear == 1)
 	<button type="button" onclick="window.location='{{ route("empresa.create") }}'" class="btn btn-info ink-reaction btn-primary addbutton" id="myBtn"><span class="fa fa-plus"></span></button>
 	@endif
-@endforeach
 
 
 
@@ -34,43 +33,40 @@ Tablas Crear Empresa
 					<th><b>Nombre</b></th>
 					<th><b>Email</b></th>
 					<th><b>Ciudad</b></th>
-					@foreach ($perfil as $itemPerfil)
-						@if ($itemPerfil->IdRol == 12)
+				
 						<th><b>Capacidades</b></th>
-						@endif
-					@endforeach
+					
 					<th><b>Funcionarios</b></th>
-					@foreach ($perfil as $itemPerfil)
-						@if ($itemPerfil->IdRol == 12)
+					
 						<th style="width: 120px;"><b>Acción</b></th>
-						@endif
-					@endforeach
+						
 				</tr>
 			</thead>
 			<tbody>
 				@foreach ($vwempresa as $empresa)
+				@if ($permiso->consultar == 1)
 				<tr>
 					<td>{{$empresa->NombreEmpresa}}</td>
 					<td>{{$empresa->Email}}</td>
 					<td>{{$empresa->Ciudad}}</td>
-					@foreach ($perfil as $itemPerfil)
+					
 						<td>
-							@if ($itemPerfil->IdRol == 12)
+						
 							<div class="col-sm-1">
 								<a href="{{route('capacidadesEmpresa.show', $empresa->IdEmpresa) }}" class="btn btn-default btn-group-xs"><span class="fa fa-gear"></span></a>
 							</div>	
-							@endif
+							
 						</td>
-					@endforeach
+					
 					<td>
 
 						<div class="col-sm-1">
 							<a href="{{route('funcionariosEmpresa.show', $empresa->IdEmpresa) }}" class="btn btn-default btn-group-xs"><span class="fa fa-users"></span></a>
 						</div>
 					</td>
-					@foreach ($perfil as $itemPerfil)
-						@if ($itemPerfil->IdRol == 12)
+					
 						<td>
+						@if ($permiso->eliminar == 1)
 							<div class="col-sm-6">
 	
 								{!! Form::open(['route' => ['empresa.destroy', $empresa->IdEmpresa], 'method' => 'DELETE']) !!}
@@ -79,15 +75,18 @@ Tablas Crear Empresa
 	
 								{!! Form::close() !!}
 							</div>
+						@endif
+						@if ($permiso->actualizar == 1)
 							<div class="col-sm-6">
 	
 								<a href="{{route('empresa.edit', $empresa->IdEmpresa) }}" class="btn btn-primary btn-block editbutton" ><div class="gui-icon"><i class="fa fa-pencil"></i></div></a>
 	
 							</div>
-						</td>
 						@endif
-					@endforeach
+						</td>
+						
 				</tr>
+				@endif
 				@endforeach
 			</tbody>
 		</table>
