+++ resources/views/programasSecad/controlProgramas/ver_tablas_crearPrograma.blade.php
@@ -14,7 +14,9 @@
 
 			{{ Breadcrumbs::render('programa') }}
 			<!-- Begin Modal -->
+			@if ($permiso->crear == 1)
 			<button type="button" onclick="window.location='{{ route("detalleprograma.index") }}'" class="btn btn-info ink-reaction btn-primary addbutton" id="myBtn"><span class="fa fa-plus"></span></button>
+			@endif
 			{{-- End modal --}}
 
 		@endsection()
@@ -35,6 +37,7 @@
 						</thead>
 						<tbody>
 							@foreach ($programas as $programa)
+							@if ($permiso->consultar == 1)
 							<tr>
 								<td>{{ $programa->Consecutivo }}</td>
 								<td>{{ $programa->Proyecto }}</td>
@@ -44,16 +47,21 @@
 									</div>
 								</td>
 								<td>
+								@if ($permiso->eliminar == 1)
 									<div class="col-sm-6">
 										{!! Form::open(['route' => ['programa.destroy', $programa->IdPrograma], 'method' => 'DELETE']) !!}
 										{!!Form::submit('x', ['class' => 'btn btn-danger deleteButton']) !!}
 										{!! Form::close() !!}
 									</div>
+									@endif
+									@if ($permiso->actualizar == 1)
 									<div class="col-sm-6">
 										<a href="{{route('programa.edit', $programa->IdPrograma) }}" class="btn btn-primary btn-block editbutton" ><div class="gui-icon"><i class="fa fa-pencil"></i></div></a>
 									</div>
+									@endif
 								</td>
 							</tr>
+							@endif
 							@endforeach
 						</tbody>
 					</table>
